last changes

+++ strtoint/strtoint.h
@@ -6,7 +6,7 @@
 /*   By: aghergut <aghergut@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 17:35:19 by aghergut          #+#    #+#             */
-/*   Updated: 2026/01/10 12:25:43 by aghergut         ###   ########.fr       */
+/*   Updated: 2026/01/10 21:04:48 by aghergut         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@
 //int to str
 char	*ft_itoa(int n);
 //str to int
-long	ft_atoi(const char *str);
+int		ft_atoi(const char *str);
 
 #endif
